// src/features/settings/types.ts
export type StoreSettings = {
  brandName: string;
  whatsappNumber: string;
  supportEmail: string;
  manualPaymentTitle: string;
  manualPaymentInstructions: string;
  bankName: string;
  accountTitle: string;
  accountNumber: string;
  iban: string;
  codEnabled: boolean;
  manualPaymentEnabled: boolean;
};

export type StoreSettingsFormValues = {
  brandName: string;
  whatsappNumber: string;
  supportEmail: string;
  manualPaymentTitle: string;
  manualPaymentInstructions: string;
  bankName: string;
  accountTitle: string;
  accountNumber: string;
  iban: string;
  codEnabled: boolean;
  manualPaymentEnabled: boolean;
};